Makese this truly autonomous.
Prior to this change, contracts created by this weren't actually autonomous.  They had an author who functioned as an owner and had limited power over the contract.  Specifically, they had the ability to terminate the proposal.  This meant that they had the power to pull the proposal at up to the last minute, or after the proposal was proposed but before it was voted on, causing disruption within the community.

With this change, the contracts generated by this factory are now truly autonomous as once they are created they cannot be mutated at all.

I have also added in the functionality to have these proposals vote no on everything *except* the proposal from this contract.  This makes it so people don't have to worry about their delegated UNI abstaining from an important proposal that is launched non-autonomously.

Recommended future additions:
Make it so there is a delay of 7 days between when the proposal reaches 10M UNI delegated and when it files the proposal and votes.  This will give ample time for people to prepare their UNI for voting on the proposal.  This is particularly important if there are multiple autonomous proposals that a person wants to support, they shouldn't have to worry about one of them going through as a surprise while they are delegated to another.