Castle V2.1
- Created a distribution strategy for the aToken interest that benefits protocol participants - LP'ers, Keepers, and the Organization.
-Tidied up the code a bit.
- Removed Mint and Burn testing functions.
- Imported one function from uniswap pair contract: Sync() - called in the distribution strategy.

PURPOSE: Creates incentives for people to participate in wrapped aTokens and liquidity pool them.

For Keepers: A percentage of the distribution is granted to the keeper for executing interest distributions to Liquidity Poolers. A future governance is intended to act as owner and set the keeper incentive rates.

For Poolers: The distribution strategy places redemption tokens in a liquidity pool determined by governance equalling the difference between the current balance of an aToken and the current total supply of wrapped aToken, less the keeper incentive.